mlflow_retrain/requirements.txt
boto3==1.28.0
pandas==1.5.3
#numpy==1.24.3
scikit-learn==1.2.2
tensorflow==2.12.0
mlflow==2.9.0
awscli==1.29.19
python-dotenv==1.0.0